requirements.txt
django==4.2.15
python-dotenv
str2bool

django-extensions
djangorestframework
channels
channels_redis
django-htmx
# UI
django-admin-soft-dashboard

# Deployment
whitenoise
gunicorn

# psycopg2-binary
# mysqlclient
django-s3-sqlite

# psycopg2
pycodestyle
psycopg2-binary

